Add template for connect to database (#3)
This file is template for connect to MySQL (PHPMyadmin) 
Processing will be used soon as sign-in completely connect and receive data from database which I DIDN'T DO FOR SIGN-IN YET 55+.
Ps.1 : There're a lot of things I didn't check it yet.  So I will check it later ASAP.
Ps.2 : You can leave textlogin.php because I just test with it.  (Not sure to test with real login.php file, 55+)